+++ auth/auth_test/auth_test.go
@@ -10,9 +10,9 @@ import (
 	os "os"
 	testing "testing"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v9"
-	client "github.com/KardFinancial/kard-go-sdk/v9/client"
-	option "github.com/KardFinancial/kard-go-sdk/v9/option"
+	kard "github.com/KardFinancial/kard-go-sdk/v10"
+	client "github.com/KardFinancial/kard-go-sdk/v10/client"
+	option "github.com/KardFinancial/kard-go-sdk/v10/option"
 	require "github.com/stretchr/testify/require"
 )