+++ auth/oauth_wire_test/oauth_wire_test.go
@@ -11,9 +11,9 @@ import (
 	strings "strings"
 	testing "testing"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v9"
-	client "github.com/KardFinancial/kard-go-sdk/v9/client"
-	option "github.com/KardFinancial/kard-go-sdk/v9/option"
+	kard "github.com/KardFinancial/kard-go-sdk/v10"
+	client "github.com/KardFinancial/kard-go-sdk/v10/client"
+	option "github.com/KardFinancial/kard-go-sdk/v10/option"
 	assert "github.com/stretchr/testify/assert"
 	require "github.com/stretchr/testify/require"
 )