+++ internal/retrier_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/KardFinancial/kard-go-sdk/v9/core"
+	"github.com/KardFinancial/kard-go-sdk/v10/core"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )