+++ users/errors.go
@@ -4,7 +4,7 @@ package users
 
 import (
 	json "encoding/json"
-	core "github.com/KardFinancial/kard-go-sdk/v9/core"
+	core "github.com/KardFinancial/kard-go-sdk/v10/core"
 )
 
 type UploadPartMultiStatus struct {