+++ users/rewards/users_rewards_test/users_rewards_test.go
@@ -10,10 +10,10 @@ import (
 	os "os"
 	testing "testing"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v9"
-	client "github.com/KardFinancial/kard-go-sdk/v9/client"
-	option "github.com/KardFinancial/kard-go-sdk/v9/option"
-	users "github.com/KardFinancial/kard-go-sdk/v9/users"
+	kard "github.com/KardFinancial/kard-go-sdk/v10"
+	client "github.com/KardFinancial/kard-go-sdk/v10/client"
+	option "github.com/KardFinancial/kard-go-sdk/v10/option"
+	users "github.com/KardFinancial/kard-go-sdk/v10/users"
 	require "github.com/stretchr/testify/require"
 )